+++ src/main/java/net/silentchaos512/torchbandolier/init/ModItems.java
@@ -18,28 +18,69 @@
 
 package net.silentchaos512.torchbandolier.init;
 
+import net.minecraft.block.Block;
+import net.minecraft.init.Blocks;
 import net.minecraft.item.Item;
+import net.minecraft.util.IItemProvider;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.silentchaos512.torchbandolier.TorchBandolier;
 import net.silentchaos512.torchbandolier.item.ItemTorchBandolier;
 
+import javax.annotation.Nullable;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.stream.Stream;
+
 public final class ModItems {
+    public static ItemTorchBandolier emptyTorchBandolier;
     public static ItemTorchBandolier torchBandolier;
+    private static final Map<Item, ItemTorchBandolier> TORCH_BANDOLIERS = new HashMap<>();
 
     private ModItems() {}
 
     public static void registerAll(RegistryEvent.Register<Item> event) {
         if (!event.getName().equals(ForgeRegistries.ITEMS.getRegistryName())) return;
 
-        torchBandolier = new ItemTorchBandolier();
-        register("torch_bandolier", torchBandolier);
+        emptyTorchBandolier = new ItemTorchBandolier((Block) null);
+        torchBandolier = new ItemTorchBandolier(Blocks.TORCH);
+        registerTorchBandolier("empty_torch_bandolier", emptyTorchBandolier);
+        registerTorchBandolier("torch_bandolier", torchBandolier);
+        registerTorchBandolier("stone_torch_bandolier", new ItemTorchBandolier(() ->
+                ForgeRegistries.BLOCKS.getValue(new ResourceLocation("silentgear:stone_torch"))
+        ));
+
+        if (TorchBandolier.isDevBuild()) {
+            registerTorchBandolier("test_item", new ItemTorchBandolier(Blocks.GLOWSTONE));
+        }
     }
 
-    private static void register(String name, Item item) {
-        ResourceLocation id = new ResourceLocation(TorchBandolier.MOD_ID, name);
-        item.setRegistryName(id);
+    @Nullable
+    public static ItemTorchBandolier getTorchBandolier(IItemProvider torch) {
+        return TORCH_BANDOLIERS.get(torch.asItem());
+    }
+
+    public static void registerTorchBandolier(ResourceLocation name, ItemTorchBandolier item) {
+        Block torch = item.getTorchBlock();
+        if (torch != null) {
+            TORCH_BANDOLIERS.put(torch.asItem(), item);
+        }
+        register(name, item);
+    }
+
+    private static void registerTorchBandolier(String name, ItemTorchBandolier item) {
+        registerTorchBandolier(TorchBandolier.getId(name), item);
+    }
+
+    private static void register(ResourceLocation name, Item item) {
+        if (item.getRegistryName() == null) {
+            item.setRegistryName(name);
+        }
         ForgeRegistries.ITEMS.register(item);
+    }
+
+    public static Stream<Map.Entry<Item, ItemTorchBandolier>> getTorchBandolierPairs() {
+        return TORCH_BANDOLIERS.entrySet().stream();
     }
 }