+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## [Unreleased]
 ### Added
+- Torch bandoliers can now support other types of torches, but these must be added as separate items (no custom types through configs or data packs)
+- Stone torch bandolier (Silent Gear)
+### Changed
+- Empty torch bandoliers are now a separate item
+- Torch type must be set by crafting a torch with an empty bandolier
+- Placing a torch bandolier in a crafting grid will allow you to extract torches, up to a stack at a time
+
+## [1.0.0]
+### Added
 - Config for max torch count
 
 ## [0.1.1] - 2018-08-19